Split pcbfoot examples apart from the module itself.
This commit is entirely made of file moves.  We don't want to install
the examples along with the module itself.  This patch moves those
examples into a separate directory from the pcbfoot library itself.

The utilities that pcbfoot provides are not, at the moment, installed
into a location where they can be used as binaries.  I think that
these should be teased apart from the library functions that they
provide.

(It's starting to look like pcbfoot should be in a separate git repo
from ods2trag to me.)